+++ lib/jbr.rb
@@ -14,10 +14,11 @@ require 'jbr/request'
 require 'jbr/oauth'
 
 require 'jbr/account'
+# Property comes before Client and Visit: their queries read its fields as they load.
+require 'jbr/property'
 require 'jbr/client'
 require 'jbr/invoice'
 require 'jbr/job'
-require 'jbr/property'
 require 'jbr/quote'
 require 'jbr/visit'